+++ neelthee_mansion/all_game_utils.py
@@ -2,13 +2,11 @@ from .utils import *
 from re import *
 
 
-def ask_for_consent(text: str) -> bool:
+def ask_for_consent(text: str, text_area) -> bool:
     while True:
-        type_text(f"{text}? Y/N")
-        anser = str(input(">")).lower().strip()
+        add_text_to_textbox(text_area, f"{text}? Y/N")
+        anser = loop_til_valid_input("", "That wasn't Y or N", Y_N).Value
         if anser == "y":
             return True
         elif anser == "n":
             return False
-        else:
-            type_text("That wasn't Y or N")